Nest secrets and projects under secrets manager
