+++ crates/bitwarden-wasm/src/client.rs
@@ -1,9 +1,9 @@
 extern crate console_error_panic_hook;
-use log::Level;
 use std::{rc::Rc, sync::RwLock};
 
 use bitwarden_json::client::Client as JsonClient;
 use js_sys::Promise;
+use log::Level;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::future_to_promise;
