+++ crates/bitwarden/src/secrets_manager/mod.rs
@@ -1,0 +1,2 @@
+pub mod projects;
+pub mod secrets;
