+++ crates/bitwarden/src/secrets_manager/projects/delete.rs
@@ -1,13 +1,14 @@
-use crate::{
-    client::Client,
-    error::{Error, Result},
-};
 use bitwarden_api_api::models::{
     BulkDeleteResponseModel, BulkDeleteResponseModelListResponseModel,
 };
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+
+use crate::{
+    client::Client,
+    error::{Error, Result},
+};
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
