+++ crates/bitwarden/src/secrets_manager/projects/get.rs
@@ -2,31 +2,31 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
+use super::ProjectResponse;
 use crate::{
+    client::Client,
     error::{Error, Result},
-    Client,
 };
-
-use super::SecretResponse;
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct SecretGetRequest {
-    /// ID of the secret to retrieve
+pub struct ProjectGetRequest {
+    /// ID of the project to retrieve
     pub id: Uuid,
 }
 
-pub(crate) async fn get_secret(
+pub(crate) async fn get_project(
     client: &mut Client,
-    input: &SecretGetRequest,
-) -> Result<SecretResponse> {
+    input: &ProjectGetRequest,
+) -> Result<ProjectResponse> {
     let config = client.get_api_configurations().await;
-    let res = bitwarden_api_api::apis::secrets_api::secrets_id_get(&config.api, input.id).await?;
+
+    let res = bitwarden_api_api::apis::projects_api::projects_id_get(&config.api, input.id).await?;
 
     let enc = client
         .get_encryption_settings()
         .as_ref()
         .ok_or(Error::VaultLocked)?;
 
-    SecretResponse::process_response(res, enc)
+    ProjectResponse::process_response(res, enc)
 }
