+++ crates/bitwarden/src/secrets_manager/projects/mod.rs
