+++ crates/bitwarden/src/secrets_manager/projects/update.rs
@@ -1,28 +1,28 @@
-use bitwarden_api_api::models::ProjectCreateRequestModel;
-
-use crate::{
-    client::Client,
-    error::{Error, Result},
-};
-
+use bitwarden_api_api::models::ProjectUpdateRequestModel;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use super::ProjectResponse;
+use crate::{
+    client::Client,
+    error::{Error, Result},
+};
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct ProjectCreateRequest {
-    /// Organization where the project will be created
+pub struct ProjectPutRequest {
+    /// ID of the project to modify
+    pub id: Uuid,
+    /// Organization ID of the project to modify
     pub organization_id: Uuid,
 
     pub name: String,
 }
 
-pub(crate) async fn create_project(
+pub(crate) async fn update_project(
     client: &mut Client,
-    input: &ProjectCreateRequest,
+    input: &ProjectPutRequest,
 ) -> Result<ProjectResponse> {
     let enc = client
         .get_encryption_settings()
@@ -31,17 +31,14 @@ pub(crate) async fn create_project(
 
     let org_id = Some(input.organization_id);
 
-    let project = Some(ProjectCreateRequestModel {
+    let project = Some(ProjectUpdateRequestModel {
         name: enc.encrypt(input.name.as_bytes(), org_id)?.to_string(),
     });
 
     let config = client.get_api_configurations().await;
-    let res = bitwarden_api_api::apis::projects_api::organizations_organization_id_projects_post(
-        &config.api,
-        input.organization_id,
-        project,
-    )
-    .await?;
+    let res =
+        bitwarden_api_api::apis::projects_api::projects_id_put(&config.api, input.id, project)
+            .await?;
 
     let enc = client
         .get_encryption_settings()
