+++ crates/bitwarden/src/secrets_manager/secrets/delete.rs
