+++ crates/bitwarden/src/secrets_manager/secrets/list.rs
