+++ crates/bitwarden/src/secrets_manager/secrets/secret_response.rs
