+++ crates/bitwarden/src/secrets_manager/secrets/update.rs
@@ -1,28 +1,30 @@
-use crate::{error::Error, error::Result, Client};
-use bitwarden_api_api::models::SecretCreateRequestModel;
+use bitwarden_api_api::models::SecretUpdateRequestModel;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use super::SecretResponse;
+use crate::{
+    client::Client,
+    error::{Error, Result},
+};
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct SecretCreateRequest {
-    /// Organization where the secret will be created
+pub struct SecretPutRequest {
+    /// ID of the secret to modify
+    pub id: Uuid,
+    /// Organization ID of the secret to modify
     pub organization_id: Uuid,
 
     pub key: String,
     pub value: String,
     pub note: String,
-
-    /// IDs of the projects that this secret will belong to
-    pub project_ids: Option<Vec<Uuid>>,
 }
 
-pub(crate) async fn create_secret(
+pub(crate) async fn update_secret(
     client: &mut Client,
-    input: &SecretCreateRequest,
+    input: &SecretPutRequest,
 ) -> Result<SecretResponse> {
     let enc = client
         .get_encryption_settings()
@@ -31,20 +33,16 @@ pub(crate) async fn create_secret(
 
     let org_id = Some(input.organization_id);
 
-    let secret = Some(SecretCreateRequestModel {
+    let secret = Some(SecretUpdateRequestModel {
         key: enc.encrypt(input.key.as_bytes(), org_id)?.to_string(),
         value: enc.encrypt(input.value.as_bytes(), org_id)?.to_string(),
         note: enc.encrypt(input.note.as_bytes(), org_id)?.to_string(),
-        project_ids: input.project_ids.clone(),
+        project_ids: None,
     });
 
     let config = client.get_api_configurations().await;
-    let res = bitwarden_api_api::apis::secrets_api::organizations_organization_id_secrets_post(
-        &config.api,
-        input.organization_id,
-        secret,
-    )
-    .await?;
+    let res =
+        bitwarden_api_api::apis::secrets_api::secrets_id_put(&config.api, input.id, secret).await?;
 
     let enc = client
         .get_encryption_settings()
